+++ src/OrderLinesList/OrderLinesList.js
@@ -55,6 +55,7 @@ const resultsFormatter = {
   updatedDate: line => <FolioFormattedDate value={get(line, 'metadata.updatedDate')} />,
   productIds: line => get(line, 'details.productIds', []).map(product => product.productId).join(', '),
   vendorRefNumber: line => get(line, 'vendorDetail.refNumber', ''),
+  title: line => get(line, 'titleOrPackage', ''),
 };
 
 export const columnMapping = {